Create SpotifyProject
The popularity of a song on Spotify is determined by the number of streams it receives, and is an important
metric for both artists and record labels. Understanding the factors that influence track popularity can be useful
for various purposes, such as guiding artists and labels in producing and promoting their music, helping
streaming platforms improve their recommendation algorithms and user experience, and enabling listeners to
discover and enjoy new music based on their preferences. This project aims to analyze the Spotify Songs
dataset to gain insights into the characteristics of the songs that make them popular among listeners. The
project will use three main data analysis approaches, word cloud analysis, linear modeling, and shiny
application, to achieve this goal.